app/requirements.txt
celery==5.0.5
coreapi==2.3.3
django==3.1.5
django-bootstrap4==2.3.1
django-celery-beat==2.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
ipython==7.19.0
psycopg2-binary==2.8.6
redis==3.5.3
requests==2.25.1
